added contiguous functions
